Make toAdaptiveColor its own extension
and rename the dir in :stats
