+++ feature/stats/src/main/java/com/github/saikcaskey/stats/data/repository/VenueRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.github.saikcaskey.pokertracker.domain.dao.VenueDao
 import com.github.saikcaskey.pokertracker.domain.datasource.UserDataSource
 import com.github.saikcaskey.pokertracker.domain.models.Venue
 import com.github.saikcaskey.pokertracker.domain.repository.VenueRepository
-import com.github.saikcaskey.stats.ext.flatMapWithUserId
+import com.github.saikcaskey.stats.extensions.flatMapWithUserId
 import kotlinx.coroutines.flow.Flow
 
 class VenueRepositoryImpl(
